update default project resources
