+++ docker-compose.yml
@@ -9,25 +9,8 @@ services:
       VIRTUAL_HOST: app-engine.subtext
       VIRTUAL_ADMIN: alonzo.turner@subtext.productions
       PHP_IDE_CONFIG: serverName=AppEngine
-      APP_CONFIG: acceptance.php
-      DB_USER: root
-      DB_PASS: password
-      DB_NAME: engine-db
-      DB_HOST: mysql
+      APP_CONFIG: unit.php
     volumes:
     - ./:/var/www
     ports:
     - 80:80
-  mysql:
-    image: mysql:8
-    container_name: app_engine_mysql
-    environment:
-      MYSQL_ROOT_PASSWORD: password
-      MYSQL_DATABASE: demo
-    volumes:
-      - "mysql8_data:/var/lib/mysql"
-    ports:
-      - 3306:3306
-
-volumes:
-  mysql8_data:
