+++ tests/unit/bootstrap.php
@@ -5,4 +5,4 @@ use Dotenv\Dotenv;
 
 $root = dirname(__DIR__, 2);
 require_once($root . '/vendor/autoload.php');
-(Dotenv::create($root))->load();
+Dotenv::createImmutable($root)->load();
